requirements.txt
fastapi
uvicorn
sqlalchemy
apscheduler
python-dotenv
requests
beautifulsoup4
feedparser
googletrans==4.0.0rc1
groq
motor
pydantic
passlib[bcrypt]
python-jose[cryptography]
python-multipart
